strategy

+++ src/theory/bags/theory_bags.cpp
@@ -287,8 +287,8 @@ void TheoryBags::postCheck(Effort effort)
 
 void TheoryBags::runStrategy(Theory::Effort e)
 {
-  std::vector<std::pair<InferStep, size_t>>::iterator it = d_strat.stepBegin(e);
-  std::vector<std::pair<InferStep, size_t>>::iterator stepEnd =
+  std::vector<std::pair<InferStep, int>>::iterator it = d_strat.stepBegin(e);
+  std::vector<std::pair<InferStep, int>>::iterator stepEnd =
       d_strat.stepEnd(e);
 
   Trace("bags-process") << "----check, next round---" << std::endl;
